refactor: extract FranchiseRecordRepository + RecordAnnouncementRepository; reduce aggregator to thin delegation shell
