+++ ibl5/classes/RecordHolders/RecordHoldersRepository.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace RecordHolders;
 
-use League\League;
 use League\LeagueContext;
 use RecordHolders\Contracts\RecordHoldersRepositoryInterface;
 
@@ -31,16 +30,18 @@ use RecordHolders\Contracts\RecordHoldersRepositoryInterface;
  */
 class RecordHoldersRepository extends \BaseMysqliRepository implements RecordHoldersRepositoryInterface
 {
-    private const ANNOUNCEMENT_CACHE_KEY = 'record_announcements_last_date';
-
     private PlayerRecordRepository $playerRecords;
     private TeamRecordRepository $teamRecords;
+    private FranchiseRecordRepository $franchiseRecords;
+    private RecordAnnouncementRepository $announcements;
 
     public function __construct(\mysqli $db, ?LeagueContext $leagueContext = null)
     {
         parent::__construct($db, $leagueContext);
         $this->playerRecords = new PlayerRecordRepository($db, $leagueContext);
         $this->teamRecords = new TeamRecordRepository($db, $leagueContext);
+        $this->franchiseRecords = new FranchiseRecordRepository($db, $leagueContext);
+        $this->announcements = new RecordAnnouncementRepository($db, $leagueContext);
     }
 
     /**
@@ -120,42 +121,7 @@ class RecordHoldersRepository extends \BaseMysqliRepository implements RecordHol
      */
     public function getMostPlayoffAppearances(): array
     {
-        $query = "SELECT
-                t.team_name,
-                COUNT(DISTINCT pr.year) AS count,
-                GROUP_CONCAT(DISTINCT pr.year ORDER BY pr.year ASC SEPARATOR ', ') AS years
-            FROM vw_playoff_series_results pr
-            JOIN `ibl_team_info` t ON t.teamid = pr.winner_tid OR t.teamid = pr.loser_tid
-            WHERE t.teamid BETWEEN 1 AND " . League::MAX_REAL_TEAMID . "
-            GROUP BY t.team_name
-            ORDER BY count DESC, t.team_name ASC
-            LIMIT 5";
-
-        $rows = $this->fetchAll($query);
-
-        // Find the max count to include ties
-        $maxCount = 0;
-        foreach ($rows as $countRow) {
-            /** @var array{team_name: string, count: int, years: string} $countRow */
-            if ($countRow['count'] > $maxCount) {
-                $maxCount = $countRow['count'];
-            }
-        }
-
-        /** @var list<PlayoffAppearanceRecord> $records */
-        $records = [];
-        foreach ($rows as $row) {
-            /** @var array{team_name: string, count: int, years: string} $row */
-            if ($row['count'] === $maxCount) {
-                $records[] = [
-                    'team_name' => $row['team_name'],
-                    'count' => $row['count'],
-                    'years' => $row['years'],
-                ];
-            }
-        }
-
-        return $records;
+        return $this->franchiseRecords->getMostPlayoffAppearances();
     }
 
     /**
@@ -198,77 +164,7 @@ class RecordHoldersRepository extends \BaseMysqliRepository implements RecordHol
      */
     public function getMostTitlesByType(string $titlePattern): array
     {
-        $rows = $this->fetchAll(
-            self::buildMostTitlesByTypeQuery(),
-            's',
-            '%' . $titlePattern . '%'
-        );
-
-        // Find the max count to include ties
-        $maxCount = 0;
-        foreach ($rows as $countRow) {
-            /** @var array{team_name: string, count: int, years: string} $countRow */
-            if ($countRow['count'] > $maxCount) {
-                $maxCount = $countRow['count'];
-            }
-        }
-
-        /** @var list<FranchiseTitleRecord> $records */
-        $records = [];
-        foreach ($rows as $row) {
-            /** @var array{team_name: string, count: int, years: string} $row */
-            if ($row['count'] === $maxCount) {
-                $records[] = [
-                    'team_name' => $row['team_name'],
-                    'count' => $row['count'],
-                    'years' => $row['years'],
-                ];
-            }
-        }
-
-        return $records;
-    }
-
-    /**
-     * Inlined team awards query with optimized Champions/HEAT branches.
-     *
-     * Uses window functions instead of correlated subqueries. The HEAT-champion
-     * branch reads from the `vw_heat_champions` view (migration 149) rather than
-     * an inline CTE.
-     */
-    private static function buildMostTitlesByTypeQuery(): string
-    {
-        return "SELECT
-                name AS team_name,
-                COUNT(*) AS count,
-                GROUP_CONCAT(year ORDER BY year ASC SEPARATOR ', ') AS years
-            FROM (
-                SELECT year, name, award
-                FROM `ibl_team_awards`
-
-                UNION ALL
-
-                SELECT ranked.year, ranked.name, 'IBL Champions' AS award
-                FROM (
-                    SELECT
-                        psr.year,
-                        psr.winner AS name,
-                        psr.round,
-                        MAX(psr.round) OVER (PARTITION BY psr.year) AS max_round,
-                        COUNT(*) OVER (PARTITION BY psr.year, psr.round) AS series_in_round
-                    FROM vw_playoff_series_results psr
-                ) ranked
-                WHERE ranked.round = ranked.max_round AND ranked.series_in_round = 1
-
-                UNION ALL
-
-                SELECT year, name, award
-                FROM `vw_heat_champions`
-            ) all_awards
-            WHERE award LIKE ?
-            GROUP BY name
-            ORDER BY count DESC, name ASC
-            LIMIT 5";
+        return $this->franchiseRecords->getMostTitlesByType($titlePattern);
     }
 
     /**
@@ -276,18 +172,7 @@ class RecordHoldersRepository extends \BaseMysqliRepository implements RecordHol
      */
     public function getLastAnnouncedDate(): ?string
     {
-        $row = $this->fetchOne(
-            "SELECT `value` FROM `cache` WHERE `cache_key` = ?",
-            's',
-            self::ANNOUNCEMENT_CACHE_KEY
-        );
-
-        if ($row === null) {
-            return null;
-        }
-
-        /** @var array{value: string} $row */
-        return $row['value'];
+        return $this->announcements->getLastAnnouncedDate();
     }
 
     /**
@@ -295,12 +180,7 @@ class RecordHoldersRepository extends \BaseMysqliRepository implements RecordHol
      */
     public function markAnnouncementsProcessed(string $gameDate): void
     {
-        $this->execute(
-            "REPLACE INTO `cache` (`cache_key`, `value`, `expiration`) VALUES (?, ?, 0)",
-            'ss',
-            self::ANNOUNCEMENT_CACHE_KEY,
-            $gameDate
-        );
+        $this->announcements->markAnnouncementsProcessed($gameDate);
     }
 
     /**
@@ -310,39 +190,6 @@ class RecordHoldersRepository extends \BaseMysqliRepository implements RecordHol
      */
     public function getUnannouncedGameDates(?string $lastAnnouncedDate): array
     {
-        // Get the latest sim's date range from `ibl_sim_dates`
-        /** @var array{start_date: string, end_date: string}|null $latestSim */
-        $latestSim = $this->fetchOne(
-            "SELECT start_date, end_date FROM `ibl_sim_dates` ORDER BY sim DESC LIMIT 1"
-        );
-
-        if ($latestSim === null) {
-            return [];
-        }
-
-        $simStart = $latestSim['start_date'];
-        $simEnd = $latestSim['end_date'];
-
-        // If the last announced date is at or after the sim end, everything is already processed
-        if ($lastAnnouncedDate !== null && $lastAnnouncedDate >= $simEnd) {
-            return [];
-        }
-
-        // Use the later of sim start or (lastAnnouncedDate + 1 day) as the floor
-        $floor = $simStart;
-        if ($lastAnnouncedDate !== null && $lastAnnouncedDate >= $simStart) {
-            $floor = $lastAnnouncedDate;
-        }
-
-        /** @var list<array{game_date: string}> $rows */
-        $rows = $this->fetchAll(
-            "SELECT DISTINCT game_date FROM `ibl_box_scores` WHERE game_date > ? AND game_date <= ? ORDER BY game_date ASC",
-            'ss',
-            $floor,
-            $simEnd
-        );
-
-        return array_map(static fn(array $row): string => $row['game_date'], $rows);
+        return $this->announcements->getUnannouncedGameDates($lastAnnouncedDate);
     }
-
 }
